Fixed bug in autocompletion
When pressing tab multiple times, the command in prompt was completed
letter by letter to display the shortest command available from first
input.
It's now fixed :
In such a case, 'mid' (the middle mark of the algorithm) was
systematically one letter higher than 'low' (lowest branch of alg)
because of the number of characters in the shortest command available
(odd number of letters).
On the other side, when the word contains a even number of letters,
'mid' is given 'low' value before exiting the while loop.
So, we can use 'low' as the mark for the left() function.