Added help tip, shortcut command & corrected bug
Help tip : when in mode "circle with center and radius", typing "radius"
in the command-line allow the user to set the radius without moving
hands off the keyboard, but the helper message didn't mentionned that
possibility.

Bug : when setting radius from commandline, pressing enter to validate
causes LibreCAD to draw a circle of selected radius at relative zero
position (if no mouse activity into the container) or crosshair
position (if mouse activity into the container).
I added a `switch` test to the `trigger()` method to test the mode, and
if the user is in mode "SetRadius", circle is not drawn & relative zero
is not moved.

Command shortcuts: `circlecr` and `cc`

Conflicts:
	librecad/src/cmd/rs_commands.cpp