Remove useless operator=() for DL_Attributes
As DL_Attributes itself does not use any dynamic memory,
the default assignment operator will do the same job.

GCC generates a warning when the default copy constructor is used
for classes that have user-defined assignment operator. This is
because such classes should implement their own copy contructor.
This is obviously not needed for DL_Attributes.