Fix issues with DL_Extrusion
DL_Extrusion used a dynamically allocated direction vector
(which is always allocated to 3 elements) without providing
a copy or move constructor. This is **obviously dangerous**.

This commit replaces the dynamically allocated vector with
a statically allocated 3 element array. This simplifies the
class as no constructor, destructor and operators have to be
provided anymore and no memory has to be manually allocated.

It also removes one of the getDirection() functions as it exposes
the internal direction pointer. Also this function isn't used
anywhere in LibreCAD, so no other code has to be modified.