Fix argument of RS_GraphicView::setBackground in export
Qt::black is an enum value and it caused RS_Color(unsigned int f) constructor to be called (but its int argument should be flags and not a color from Qt's enum). So basically the meaning of the int value was changed.

This change surely fixes this problem:
- draw a black line
- export as PNG
- Background: white
- Colouring: coloured
- problem: the exported PNG does not contain the line (it is blank)

This change may cause a fix in bugs #962 and #553 (or at least solve some confusion in these bugs).