Fix renaming of nested blocks
Block list fix:
When block A that is nested within block B is renamed, then rename
also all inserts of block A withing block B in the blocklist.

Entity container fix:
RS_Insert is a subclass of RS_EntityContainer. Inserts within
inserts (in entity tree of a graphic) should be renamed as well.

Fixes issue #1527