Issue #1515 Don't scroll blockwidget after toggling block visibility
Before the change (as noted by
https://github.com/LibreCAD/LibreCAD/issues/1515#issuecomment-1051356059),
toggling block visibility caused it to scroll to the active
block. Such behaviour is problematic especially for drawings with
large number of blocks.

It looks like blockwidget code assumed that QG_BlockWidget::update()
would be called after every user interaction to maintain current
scroll position. It turns out that toggling block visibility bypassed
the update method - it was handled by slotActivated() and
activateBlock() later on. To fix the problem, the activateBlock()
method now also restores the scroll position. As it is called by
update(), the scroll position handling in update() seems redundant and
has been removed.