issue #1616: block copy/paste avoid double rotation
this is a quick fix only, probably should be replaced with an actual fix
using a block reference definition.

By this quick fix, a rotated block has entities rotated and a rotation
angle, but a copy of this rotated block will have rotated entities with
a zero rotation angle. This means the block and its copy may have
difference more than a translation.