Fixed a crash from edit property
Steps to reproduce the crash:

1, Draw a circle(or line, etc.);
2, Double click on the circle to edit its properties;
3, Change its radius (or center), and OK;
4, undo.

The crash is due to editing of a temporary entity created for
highlighting. The temporary entity is removed automatically later on,
and out of any undo/redo cycle, but the modify action would record the
temporary entity in an undo cycle, causing any subsequent redo/undo
to use a deleted entity.

This quick fix, always use the real entity for modifying.

Long term thought, keep the temporary entities by shared_ptr.