Issue #1653: closest point on ellipse robustness
The existing analytical solution is not robust for two cases:
1. small eccentricity;
2. the given point is close to the origin of the ellipse.

The fix:

Use Newton-Raphson for an approximate solution when the equation
parameters are malformed;
It's okay to accept poor precision of the
solution in those cases. For example, when the given point is at the
ellipse origin, there are actually two closest points for an ellipse.
If the ellipse is degenerate as a circle, every point on the circle
should be a solution.