feat: Shortcut "." to specify current point coordinates on command line
In using the LibreCAD command line, I find that far and away the most
common point that needs to be entered is simply the current point (the
origin for relative coordinates). I am aware that "0..0" specifies this,
but that is four fairly awkward keystrokes. Hence, I suggest that a
single-character shortcut is well worth it. This commit implements "."
to mean the current point (i.e. same as "0..0") when coordinates are
expected on the command line. The shortcut was chosen because a bare
"." has no other meaning, and it could be read "point" anyway, as a
convenient mnemonic. Happy to revise in favor of any other single
character if desired.