Command widget: add method to re-dock
Before this commit, it's easy to make the command widget float. Either
by the "Floating" button on the title bar, or by dragging the command
widget. However, there's not straightforward to re-dock a widows after
the window is floating.

This commit added a button to allow easy re-docking a floating command
widget.