Issue #1753 : toggle snapping by spacebar
The spacebar was used to activate the command widget otherwise.

If (Options -> Application Preferences -> Defaults -> " Toggle free snap
mode when SPACE BAR is pressed") is set, spacebar won't trigger activate
the command line (use Ctl-M,or Key_Colon ':' to activate the command
line).

If the option is not set, pressing spacebar would activate the command
widget.

As other application preferences, need to restart the application after
this option change.