Issue #1777: scaling factor by points
Issue #1777:

Experimental feature: for the scaling tool, allow to specify a source
point and a target point, so the scaling would scale the source to the
target point.

If the coordinate difference from the scaling center is zero in any
direction(x- or y-), no scaling is applied in that direction.

Negative factors are allowed, and this may not work well with all entity
types.