Issue #523: recover tangent points
The algorithm: Around a tangent point, the two parent curves should have a single intersection. However, due to rounding errors the rounding errors may cause the intersection find algorithms to find 0, 1, or 2 intersections in this case.

The basic idea: slightly offset a curve to allow finding the intersection reliably.

1. Check offsets of two sides;
2. If two intersections are found, reduce the distance to the offset,
and bisection is used to find the tangent point.