Issue #2012: lock relative zero logic
The issue: locked relative zero got unlocked by new actions

The root cause: in EventHandler::setQAction() would save the QAction in
the handler, and the handler would uncheck the QAction, after it's
considered to be finished. This is clearly wrong for the case of Locked
Relative Zero.

The quick fix: the QActions are not required to be handled by the
EventHandler. Avoid sending the QActions to the EventHandler in
QG_GraphicView::setCurrentQAction(QAction* q_action).

The fix is experimental: if some QActions need to be managed by the
EventHandler, this fix would cause regressions for those.