Issue #2088: solid fill regression
This is Work-in-Progress, and the fix is not complete yet.

1. Added a new algorithm (LoopOptimizer) for loop analysis;
LoopOptimizer takes contour edges, and output closed contours, with each
contour ordered by entity start/end points. It means each output contour
has edges in a cyclic order, with the end point of the previous entity
is coincident with the start point of the current entity;
2. LoopOptimizer is based on R-Tree, so quadratic complexity in
RS_EntityContainer::optimizeContours() is avoided;
3. Not enough error reporting yet: not closed contours,
self-intersecting, etc;
4. Elliptic arc support is still buggy.