Command: fixed an error in alias File handling
The issue: in updateAlias(), a QFile is used for both reading and
updating the alias file.The QFile.open() in updating the alias file may
fail, since the file is being opened.

The fix: use a QString for the alias file in updateAlias(), and QFile
constructed and destroyed in reading and updating the alias file.