Issue #2128 : more RS_EntityContainer copy/move fixes
More work needed to cleanup ownership of RS_EntityContainer

Currently:

If the container has ownership, a deep copy is performed for childrens
at copy ctor/assigner. Or only a shallow copying of child pointers are
copied into the container.

A code review is needed on the correctness of the algorithm.

It's also possible to make all children as shared_ptr to avoid manually
handling child lifetimes.